requirements.txt
django==5.1.6
django-debug-toolbar==5.0.1
Pillow==11.1.0
environs[django]==14.1.1
django-phonenumber-field[phonenumbers]==8.0.0
djangorestframework==3.15.2
